requirements.txt
google-genai
pydantic
boto3
